LPD-69822 playwright: Updates for readability
