+++ modules/test/playwright/tests/search-experiences-web/main/sxpBlueprint.spec.ts
@@ -651,167 +651,3 @@ test.describe('Blueprint can be registered as a collection provider', () => {
 		});
 	});
 });
-
-test.describe('View links to Liferay Documentation', () => {
-	test('Blueprint and element editor have working links @LPS-147066', async ({
-		apiHelpers,
-		context,
-		editSXPBlueprintPage,
-		page,
-		sxpBlueprintsAndElementsViewPage,
-	}) => {
-		const sxpBlueprint =
-			await apiHelpers.searchExperiences.createSXPBlueprint();
-
-		const assertNewPageTitle = async (title: string) => {
-			const pagePromise = context.waitForEvent('page');
-
-			const newTab = await pagePromise;
-
-			await newTab.waitForLoadState();
-
-			await expect(newTab).toHaveTitle(new RegExp(title));
-
-			await newTab.close();
-		};
-
-		await test.step('Navigate to created blueprint', async () => {
-			await sxpBlueprintsAndElementsViewPage.goto();
-
-			await sxpBlueprintsAndElementsViewPage.selectTableLink(
-				sxpBlueprint.title
-			);
-		});
-
-		await test.step('Check sidebar link', async () => {
-			await editSXPBlueprintPage.addSXPElementSidebar
-				.getByRole('link', {name: 'Learn more. (Opens a new'})
-				.click();
-
-			await assertNewPageTitle('Search Blueprints Elements Reference');
-		});
-
-		await test.step('Check query settings - indexer clauses', async () => {
-			await editSXPBlueprintPage.goToQuerySettingsMenuItem();
-
-			await page
-				.getByRole('button', {name: 'Search Framework Indexer'})
-				.locator('.lexicon-icon-question-circle')
-				.click();
-
-			await expect(editSXPBlueprintPage.infoSidebar).toBeVisible();
-
-			await editSXPBlueprintPage.infoSidebar
-				.getByRole('link', {
-					name: 'Learn more. (Opens a new',
-				})
-				.click();
-
-			await assertNewPageTitle('Creating and Managing Search Blueprints');
-		});
-
-		await test.step('Check query settings - query contributors', async () => {
-			await page
-				.getByRole('button', {
-					name: 'Search Framework Query Contributors',
-				})
-				.locator('.lexicon-icon-question-circle')
-				.click();
-
-			await expect(editSXPBlueprintPage.infoSidebar).toBeVisible();
-
-			await editSXPBlueprintPage.infoSidebar
-				.getByRole('link', {
-					name: 'Learn more. (Opens a new',
-				})
-				.click();
-
-			await assertNewPageTitle('Creating and Managing Search Blueprints');
-		});
-
-		await test.step('Check Configuration tab', async () => {
-			await editSXPBlueprintPage.goToConfigurationTab();
-
-			const configurationTexts = [
-				'Enter additional blueprints',
-				'Add aggregations to those',
-				'Override the default search',
-				'Add sorts to those already',
-				'Declare new template',
-				'Add source includes and',
-			];
-			for (const text of configurationTexts) {
-				await page
-					.locator('.sheet-text', {
-						has: page.getByText(new RegExp(text)),
-					})
-					.locator('.learn-message')
-					.click();
-
-				await assertNewPageTitle(
-					'Search Blueprints Configuration Reference'
-				);
-			}
-		});
-
-		await test.step('Go to elements', async () => {
-			await editSXPBlueprintPage.cancelBlueprint();
-
-			await sxpBlueprintsAndElementsViewPage.goToElementsTab();
-
-			await sxpBlueprintsAndElementsViewPage.selectTableLink(
-				'Text Match Over Multiple Fields'
-			);
-		});
-
-		await test.step('Check info sidebar', async () => {
-			await page.getByLabel('Info').click();
-
-			await editSXPBlueprintPage.infoSidebar
-				.getByRole('link', {
-					name: 'Learn more. (Opens a new',
-				})
-				.click();
-
-			await assertNewPageTitle('Creating Elements');
-		});
-	});
-
-	test('Blueprint options portlet has working link @LPS-147066', async ({
-		context,
-		layout,
-		page,
-		searchPage,
-	}) => {
-		await test.step('Add search blueprint options to new page', async () => {
-			await page.goto('/web/guest' + layout.friendlyURL);
-
-			await searchPage.addPortlet('Blueprints Options', 'Search');
-		});
-
-		await test.step('Open blueprint options configuration modal', async () => {
-			await page
-				.getByRole('link', {name: 'Configure blueprints options'})
-				.click();
-		});
-
-		await test.step('Check blueprint link for search page', async () => {
-			await page
-				.frameLocator('iframe[id="modalIframe"]')
-				.getByRole('link', {
-					name: 'Learn how to use blueprints',
-				})
-				.click();
-
-			const pagePromise = context.waitForEvent('page');
-
-			const newTab = await pagePromise;
-
-			await newTab.waitForLoadState();
-
-			await expect(newTab).toHaveTitle(
-				/Using a Search Blueprint on a Search Page/
-			);
-		});
-	});
-});
